Add argument to specify additional archs to consider (#5)
The YumBase object's list of considered archs only includes archs that are compatible with the host arch. For instance, on an x86_64 host, ARM architectures aren't on this list. The consequence of this is that packages for archs that are not in this list are not considered.

In particular, this means it's currently not possible to manage an ARM repo from an x86_64 host, as the existing list of ARM packages would be ignored.

Add an --archs option that allows supplying a list of archs that should be considered for the repo operations.